Use package for galaxy-lib dependencies
The manual installation of dependencies introduced in
https://github.com/AndreasSko/docker-galaxy-stable/pull/27
is a bit hacky and may become broken if the file/directory
layout of Galaxy changes in the future. We now use the
package
`galaxy-job-execution` to install the galaxy-lib
and galaxy-ext dependencies directly into the venv,
which is also accessible from the HTCondor and Slurm
clusters.